Refactor AccountLink by breaking out code to concerns
Move out code for OAuth, Gitlab and Github from `AccountLink`
to three concerns responsible for each part.

For now, Gitlab and Gitlab requires the OAuthClient concern to
make the API calls, leaving it up to the using class to include this
concern. It would be nice if the Gitlab and Github concerns could
include this themselves.

Fixes #17
