+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-*Nothing yet*
+- Log in using your Gitlab or Github account.
+- Import projects from Gitlab or Github.
+- Admin area with list of registered users.
 
 [Unreleased]: https://github.com/ephracis/appatite/compare/8c736d6...HEAD
